// types/car/car.ts
export type CarData = {
  id: string;
  model: string;
  brand: string;
  year: number;
  color: string;
  mileage: number;
  identificationName: string;
  engine: string;
  fuelType: string;
  transmission: string;
  licensePlate: string;
  purchaseDate: string;
  image: string;
};

export type CarOption = {
  id: number;
  name: string;
};